demand_tangent_tuple moved out of person.py
